// apps/web/src/data/mock-data.d.ts
import type { SearchResultGroup } from '@xala/ds';
import type { UiListing } from '@xala/sdk';
export declare const demoSearchResults: SearchResultGroup[];
export declare const listingTypeOptions: {
    id: string;
    label: string;
}[];
export declare const capacityOptions: {
    id: string;
    label: string;
    min: number;
    max: number;
}[];
export declare const mockListings: UiListing[];